requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file requirements.txt requirements.in
#
certifi==2017.7.27.1      # via requests
chardet==3.0.4            # via requests
click==6.7
elasticsearch==5.4.0
idna==2.6                 # via requests
pyyaml==3.12
requests==2.18.4
six==1.11.0               # via structlog
structlog==17.2.0
urllib3==1.22             # via elasticsearch, requests
